string overflow

+++ kernel/string/itoa.c
@@ -7,7 +7,7 @@
 
 #include <types.h>
 
-void uitoa3(uint64_t n, char *s, uint8_t len, uint8_t base){
+void uitoa3(uint64_t n, char *s, uint16_t len, uint8_t base){
 	uint8_t i = 0;
 	char *rev = s;
 	char tmp;
